requirements.txt
requests~=2.32.3
websockets~=14.1
fastapi~=0.115.6
uvicorn~=0.34.0
python-dotenv~=1.0.1
pydantic~=2.10.4
pyjwt~=2.10.1